requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
jinja2==3.1.4
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
